% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.create.johnson.su.R
\name{dist.create.johnson.su}
\alias{dist.create.johnson.su}
\title{Create Distribution - Johnson SU}
\usage{
dist.create.johnson.su(
  parms = list(gamma = NA, delta = NA, xi = NA, lambda = NA, type = "SU"),
  gamma = NA,
  delta = NA,
  xi = NA,
  lambda = NA
)
}
\arguments{
\item{parms}{parameters, likely from SuppDists JohnsonFit function}

\item{gamma}{gamma parameter}

\item{delta}{delta parameter}

\item{xi}{xi parameter}

\item{lambda}{lambda parameter.}
}
\value{
A data frame with functions, distribution name, and parameters.
}
\description{
Create distribution, cumulative distribution, quantile, and random generation functions for a distribution. 
Use parameters from JohnsonFit function in SuppDists, recommend.johnson in lolcat, or specify parameters directly.
}
