% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.fit.johnson.su.R
\name{dist.fit.johnson.su}
\alias{dist.fit.johnson.su}
\title{Fit Distribution - Johnson SU}
\usage{
dist.fit.johnson.su(x, probs = c(0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{x}{Vector - data to be fit}

\item{probs}{Vector - quantiles to use for fitdistrplus (method=qme)}
}
\value{
A list with distribution cumulative distribution, quantile, and random generation functions and fit details.
}
\description{
Fit distribution, cumulative distribution, quantile, and random generation functions for a distribution.
}
