% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolcatext.R
\docType{package}
\name{lolcatext}
\alias{lolcatext}
\title{lolcatext: An extension for lolcat providing external dependencies}
\description{
The lolcatext package provides external dependency functionality not provided in the core lolcat package.
}
